Self-cleaning obsolete disks in FreeBSD plugin
