+++ CMakeLists.txt
@@ -31,6 +31,7 @@ set(NETDATA_SOURCE_FILES
         src/freebsd_sysctl.c
         src/freebsd_getmntinfo.c
         src/freebsd_getifaddrs.c
+        src/freebsd_devstat.c
         src/global_statistics.c
         src/global_statistics.h
         src/health.c
